pyproject.toml
[project]
name = "resume-bot"
version = "0.1.0"
description = "飞书简历库 Bot - 基于飞书 Bot 的智能简历管理系统"
requires-python = ">=3.11"
dependencies = [
    "lark-oapi>=1.5.0",
    "openai>=1.0.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = []
